feat: serve spacelift_stack_dependency from the Framework provider
Registers the resource in provider_framework.go and drops it from the SDKv2
ResourcesMap in the same commit, since having it in both makes the mux fail at
startup with a duplicate-resource error. The SDKv2 implementation is deleted and the
_framework.go file renamed over it; parseStackDependencyID moves across with it, as
nothing outside the file used it.

Both acceptance test files move from testSteps to testStepsMux. Their configs stand
up two spacelift_stack resources to hang the dependency off, and those are still
SDKv2, so a Framework-only factory can't serve them.

Also adds TestMuxedProviderSchema, which resolves the muxed schema the way Terraform
does on startup. Double registration is a runtime error, so go build misses it and
the acceptance tests that would trip over it need TF_ACC and credentials. This runs
in plain CI and will fail the moment a future migration forgets one half of the swap.
