+++ spacelift/provider_framework.go
@@ -124,7 +124,9 @@ func (p *frameworkProvider) Configure(ctx context.Context, req fwprovider.Config
 }
 
 func (p *frameworkProvider) Resources(_ context.Context) []func() resource.Resource {
-	return []func() resource.Resource{}
+	return []func() resource.Resource{
+		NewStackDependencyResource,
+	}
 }
 
 func (p *frameworkProvider) DataSources(_ context.Context) []func() datasource.DataSource {
